requirements.txt
Django==5.1.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
pillow==11.0.0
ruff==0.7.1
